backend/requirements.txt
flask
flask-cors
firebase-admin
pymongo
python-dotenv
